refund,release product

+++ Khoaluan/Khoaluan/InterfacesService/IUserService.cs
@@ -5,7 +5,7 @@ namespace Khoaluan.InterfacesService
 {
     public interface IUserService
     {
-        Users updateBalance(int userID, int price, int type);
+        Users updateBalance(int userID, decimal price, int type);
         int SignIn(LoginViewModel model);
         int SignUp(RegisterViewModel model);
         int ChangePassword(ChangePasswordViewModel model, int accountId);
